requirements.txt
# PIM Dashboard Requirements
# Install with: pip3.10 install --user -r requirements.txt

# Core Flask Framework
Flask==3.0.3
Flask-SocketIO==5.4.0

# HTTP Requests and Web Scraping
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0

# Google Sheets Integration
gspread==6.1.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Data Processing
pandas==2.2.2
numpy==1.26.4  # Stick to 1.26.4 for pandas compatibility

# Environment Variables
python-dotenv==1.0.1

# Async and Threading Support
python-socketio==5.11.3
python-engineio==4.9.1

# HTTP Server (for production)
gunicorn==22.0.0

# Optional: For better JSON handling
simplejson==3.19.3

# Optional: For data validation
cerberus==1.3.5

# Optional: For better logging
colorlog==6.8.2
